Refactors non-rest translator methods into translator_utils.
